requirements/prod.txt
django==5.2.4
djangorestframework==3.16.0
django-environ==0.12.0
faststream[rabbit]==0.5.48
phonenumbers==9.0.10
pika==1.3.2
redis==6.2.0
